requirements.txt
websockets==12.0
python-dotenv==1.0.0
pydantic==2.5.2
loguru==0.7.2
pyham-pe==0.1.0 